Separating single JS-file to modules
