+++ src/view/board-container.js
@@ -1,0 +1,1 @@
+export const createBoardContainer = () => `<section class="board container"></section>`;
