+++ src/view/board-task-container.js
@@ -1,0 +1,1 @@
+export const createBoardTasksContainer = () => `<div class="board__tasks"></div>`;
