+++ src/view/task-card.js
@@ -1,0 +1,36 @@
+export const createTaskCard = () => {
+  return `<article class="card card--black">
+    <div class="card__form">
+      <div class="card__inner">
+        <div class="card__control">
+          <button type="button" class="card__btn card__btn--edit">
+            edit
+          </button>
+          <button type="button" class="card__btn card__btn--archive">
+            archive
+          </button>
+          <button
+            type="button"
+            class="card__btn card__btn--favorites"
+          >
+            favorites
+          </button>
+        </div>
+
+        <div class="card__color-bar">
+          <svg class="card__color-bar-wave" width="100%" height="10">
+            <use xlink:href="#wave"></use>
+          </svg>
+        </div>
+
+        <div class="card__textarea-wrap">
+          <p class="card__text">Example task without date.</p>
+        </div>
+
+        <div class="card__settings">
+          <div class="card__details"></div>
+        </div>
+      </div>
+    </div>
+  </article>`;
+};
